Fixed all warnings from tests.

+++ tests/check_slug.c
@@ -63,7 +63,7 @@ START_TEST (test_inside_slug)
     int slug_len = 0;
     char * pattern = "/user/{name:\\s+}/to/{id}";
     char * offset = strchr(pattern, '{') + 2;
-    ck_assert( inside_slug(pattern, strlen(pattern), offset) );
+    ck_assert( (int)inside_slug(pattern, strlen(pattern), offset) );
     ck_assert( *(inside_slug(pattern, strlen(pattern), offset)) == '{' );
     ck_assert( ! inside_slug(pattern, strlen(pattern), pattern) );
 }
